tsql-core/src/test/java/com/xy/xsql/tsql/converter/functions/datatype/IdentIncrConverterTest.java
package com.xy.xsql.tsql.converter.functions.datatype;

import com.xy.xsql.block.core.printer.MetaBlockPrinter;
import com.xy.xsql.block.model.BlockMeta;
import org.junit.Assert;
import org.junit.Test;

import java.io.StringWriter;

import static org.junit.Assert.*;

/**
 * Created by xiaoyao9184 on 2018/5/31.
 */
public class IdentIncrConverterTest {

    @Test
    public void testMetaPrint() throws Exception {
        BlockMeta b = IdentIncrConverter.meta;

        StringWriter writer = MetaBlockPrinter.print(b);

        System.out.println(writer);
        Assert.assertEquals(writer.toString(),
                "<IDENT_INCR> ::=\n" +
                        "IDENT_INCR ( 'table_or_view' )");
    }

}